:recycle: Move gen func
